# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NFT Marketplace API",
    "version": "1.0.0",
    "description": "API for managing NFTs in Notion database"
  },
  "servers": [
    {
      "url": "https://nft-marketplace-iurii.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error description"
          }
        }
      },
...
      "CartData": {
        "type": "object",
        "properties": {
          "page_id": {
            "type": "string",
            "description": "Unique Notion page ID"
          },
          "title": {
            "type": "string",
            "description": "Cart title"
          },
          "user_id": {
            "type": "string",
            "description": "User ID who owns the cart"
          },
          "nfts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of NFT addresses"
          },
          "created_time": {
            "type": "string",
            "format": "date-time",
            "description": "Creation time"
          },
          "last_edited_time": {
            "type": "string",
            "format": "date-time",
            "description": "Last modification time"
          }
        }
      },
      "AddToCartRequest": {
        "type": "object",
        "properties": {
          "nft_address": {
            "type": "string",
            "description": "NFT address to add to cart"
          }
        },
        "required": [
          "nft_address"
        ]
      },
      "RemoveFromCartRequest": {
        "type": "object",
        "properties": {
          "nft_address": {
            "type": "string",
            "description": "NFT address to remove from cart"
          }
        },
        "required": [
          "nft_address"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/cart/{userid}": {
      "get": {
        "summary": "Get cart by user ID",
        "description": "Returns a specific cart by user ID",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userid",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Cart found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CartData"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cart not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add NFT to cart",
        "description": "Adds NFT to user's cart. Creates cart if it doesn't exist.",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userid",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NFT added to cart successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Operation success status"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CartData"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Cart created and NFT added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Operation success status"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CartData"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or NFT already in cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove NFT from cart",
        "description": "Removes specific NFT from user's cart",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userid",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveFromCartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NFT removed from cart successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Operation success status"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CartData"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, cart not found or NFT not in cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Cart not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}